requirements.txt
openai>=1.0.0
python-telegram-bot>=21.0
apscheduler>=3.10.0
feedparser>=6.0.0
pyyaml>=6.0
youtube-transcript-api>=1.0.0
requests>=2.31.0
python-slugify>=8.0.0
trafilatura>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
